+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## CHANGES
 
  - [BC-BREAK] Reflector parameter `$method` is no longer nullable. Only `ReflectionMethod` should be constructed.
+ - [BC-BREAK] Default values in parameters are not resolved.
 
 ### 0.14.0