+++ CHANGES.md
@@ -2,6 +2,7 @@
 
  - [BC-BREAK] Reflector parameter `$method` is no longer nullable. Only `ReflectionMethod` should be constructed.
  - [BC-BREAK] Default values in parameters are not resolved.
+ - [BC-BREAK] Given `$make` function doesn't retrieve dependencies, just make object without cache.
 
 ### 0.14.0